+++ src/lib/initKoop.js
@@ -1,0 +1,17 @@
+const config = require('config')
+const Koop = require('koop')
+
+module.exports = function () {
+  const koop = new Koop(config)
+
+  if (config.cache !== 'local') {
+    const cache = require('koop-pgcache')
+    koop.register(cache)
+  }
+
+  if (config.filesystem.s3 && config.filesystem.s3.bucket) {
+    const fs = require('koop-s3fs')
+    koop.register(fs)
+  }
+  return koop
+}