+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Changed
+* Share code for initializing Koop between `copyFile` and `exportFile`
+
 ## [1.3.0] - 2017-01-31
 ### Changed
 * Compatible with Koop 3.x latest